+++ CHANGELOG.md
@@ -93,6 +93,7 @@
 * ffi(nostr): expose `Kind::is_*` methods ([Yuki Kishimoto])
 * ffi(sdk): expose `MockRelay` ([Yuki Kishimoto])
 * js(nostr): add `Kind` object ([Yuki Kishimoto])
+* js(nostr): expose `getNip05Profile` function ([Yuki Kishimoto])
 * book: add some python examples ([RydalWater])
 
 ### Fixed